+++ src/EditorFeatures/Core.Wpf/Interactive/InteractiveEvaluator.cs
@@ -3,27 +3,19 @@
 // See the LICENSE file in the project root for more information.
 
 extern alias InteractiveHost;
-extern alias Scripting;
 
 using System;
 using System.Collections.Generic;
 using System.Collections.Immutable;
 using System.Diagnostics;
-using System.Globalization;
 using System.IO;
-using System.Linq;
-using System.Runtime.CompilerServices;
-using System.Threading;
 using System.Threading.Tasks;
 using Microsoft.CodeAnalysis.Editor.Implementation.Interactive;
 using Microsoft.CodeAnalysis.Editor.Shared.Extensions;
 using Microsoft.CodeAnalysis.Editor.Shared.Utilities;
 using Microsoft.CodeAnalysis.ErrorReporting;
 using Microsoft.CodeAnalysis.Host;
-using Microsoft.CodeAnalysis.Scripting.Hosting;
-using Microsoft.CodeAnalysis.Shared.Extensions;
 using Microsoft.CodeAnalysis.Shared.TestHooks;
-using Microsoft.CodeAnalysis.Text;
 using Microsoft.VisualStudio.InteractiveWindow;
 using Microsoft.VisualStudio.InteractiveWindow.Commands;
 using Microsoft.VisualStudio.Text;
@@ -34,7 +26,6 @@ using Roslyn.Utilities;
 namespace Microsoft.CodeAnalysis.Editor.Interactive
 {
     using InteractiveHost::Microsoft.CodeAnalysis.Interactive;
-    using RelativePathResolver = Scripting::Microsoft.CodeAnalysis.RelativePathResolver;
 
     internal sealed class InteractiveEvaluator : IResettableInteractiveEvaluator
     {
@@ -42,16 +33,12 @@ namespace Microsoft.CodeAnalysis.Editor.Interactive
 
         private readonly InteractiveEvaluatorLanguageInfoProvider _languageInfo;
 
-        private readonly InteractiveHost _interactiveHost;
-
-        private readonly string _hostDirectory;
-
         private readonly IThreadingContext _threadingContext;
-        private readonly IContentType _contentType;
         private readonly IViewClassifierAggregatorService _classifierAggregator;
         private readonly IInteractiveWindowCommandsFactory _commandsFactory;
         private readonly ImmutableArray<IInteractiveWindowCommand> _commands;
-        private readonly CancellationTokenSource _shutdownCancellationSource;
+        private readonly InteractiveWorkspace _workspace;
+        private readonly InteractiveSession _session;
 
         private IInteractiveWindow? _lazyInteractiveWindow;
         private IInteractiveWindowCommands? _lazyInteractiveCommands;
@@ -67,38 +54,10 @@ namespace Microsoft.CodeAnalysis.Editor.Interactive
 
         #endregion
 
-        #region State only accessible by queued tasks
+        public IContentType ContentType { get; }
 
-        // Use to serialize InteractiveHost process initialization and code execution.
-        // The process may restart any time and we need to react to that by clearing 
-        // the current solution and setting up the first submission project. 
-        // At the same time a code submission might be in progress.
-        // If we left these operations run in parallel we might get into a state
-        // inconsistent with the state of the host.
-        private readonly TaskQueue _taskQueue;
-
-        private readonly InteractiveWorkspace _workspace;
-        private ProjectId? _lastSuccessfulSubmissionProjectId;
-        private ProjectId? _currentSubmissionProjectId;
-        public int SubmissionCount { get; private set; }
-
-        private RemoteInitializationResult? _initializationResult;
-        private InteractiveHostPlatformInfo _platformInfo;
-        private ImmutableArray<string> _referenceSearchPaths;
-        private ImmutableArray<string> _sourceSearchPaths;
-        private string _workingDirectory;
-
-        /// <summary>
-        /// Buffers that need to be associated with a submission project once the process initialization completes.
-        /// </summary>
-        private readonly List<(ITextBuffer buffer, string name)> _pendingBuffers = new List<(ITextBuffer, string)>();
-
-        #endregion
-
-        internal InteractiveEvaluatorResetOptions ResetOptions { get; set; }
+        public InteractiveEvaluatorResetOptions ResetOptions { get; set; }
             = new InteractiveEvaluatorResetOptions(InteractiveHostPlatform.Desktop64);
-
-        InteractiveEvaluatorResetOptions IResettableInteractiveEvaluator.ResetOptions { get => ResetOptions; set => ResetOptions = value; }
 
         internal InteractiveEvaluator(
             IThreadingContext threadingContext,
@@ -114,27 +73,40 @@ namespace Microsoft.CodeAnalysis.Editor.Interactive
             Debug.Assert(languageInfo.InteractiveResponseFileName.IndexOfAny(new[] { Path.DirectorySeparatorChar, Path.AltDirectorySeparatorChar }) == -1);
 
             _threadingContext = threadingContext;
-            _taskQueue = new TaskQueue(listener, TaskScheduler.Default);
-            _shutdownCancellationSource = new CancellationTokenSource();
-            _contentType = contentType;
+            ContentType = contentType;
             _languageInfo = languageInfo;
-            _workspace = new InteractiveWorkspace(hostServices);
             _classifierAggregator = classifierAggregator;
             _commandsFactory = commandsFactory;
             _commands = commands;
-            _hostDirectory = Path.Combine(Path.GetDirectoryName(typeof(InteractiveEvaluator).Assembly.Location), "InteractiveHost");
 
-            // The following settings will apply when the REPL starts without .rsp file.
-            // They are discarded once the REPL is reset.
-            _referenceSearchPaths = ImmutableArray<string>.Empty;
-            _sourceSearchPaths = ImmutableArray<string>.Empty;
-            _workingDirectory = initialWorkingDirectory;
+            _workspace = new InteractiveWorkspace(hostServices);
 
-            _interactiveHost = new InteractiveHost(languageInfo.ReplServiceProviderType, initialWorkingDirectory);
-            _interactiveHost.ProcessInitialized += ProcessInitialized;
+            _session = new InteractiveSession(_workspace, threadingContext, listener, languageInfo, initialWorkingDirectory);
+            _session.Host.ProcessInitialized += ProcessInitialized;
         }
 
-        public IContentType ContentType => _contentType;
+        public void Dispose()
+        {
+            _session.Host.ProcessInitialized -= ProcessInitialized;
+
+            _session.Dispose();
+            _workspace.Dispose();
+
+            var interactiveWindow = _lazyInteractiveWindow;
+            if (interactiveWindow != null)
+            {
+                interactiveWindow.SubmissionBufferAdded -= SubmissionBufferAdded;
+            }
+        }
+
+        private void ProcessInitialized(InteractiveHostPlatformInfo platformInfo, InteractiveHostOptions options, RemoteExecutionResult result)
+        {
+            _ = _threadingContext.JoinableTaskFactory.RunAsync(async () =>
+            {
+                await _threadingContext.JoinableTaskFactory.SwitchToMainThreadAsync();
+                CaptureClassificationSpans();
+            });
+        }
 
         public IInteractiveWindow? CurrentWindow
         {
@@ -152,7 +124,7 @@ namespace Microsoft.CodeAnalysis.Editor.Interactive
                 _lazyInteractiveWindow = value ?? throw new ArgumentNullException(nameof(value));
                 _workspace.Window = value;
 
-                Task.Run(() => _interactiveHost.SetOutputs(value.OutputWriter, value.ErrorOutputWriter));
+                Task.Run(() => _session.Host.SetOutputs(value.OutputWriter, value.ErrorOutputWriter));
 
                 value.SubmissionBufferAdded += SubmissionBufferAdded;
                 _lazyInteractiveCommands = _commandsFactory.CreateInteractiveCommands(value, CommandPrefix, _commands);
@@ -164,7 +136,8 @@ namespace Microsoft.CodeAnalysis.Editor.Interactive
         /// </summary>
         public event Action<InteractiveHostPlatform>? OnBeforeReset;
 
-        #region Initialization
+        public int SubmissionCount
+            => _session.SubmissionCount;
 
         private IInteractiveWindow GetInteractiveWindow()
             => _lazyInteractiveWindow ?? throw new InvalidOperationException(EditorFeaturesResources.Engine_must_be_attached_to_an_Interactive_Window);
@@ -172,19 +145,13 @@ namespace Microsoft.CodeAnalysis.Editor.Interactive
         private IInteractiveWindowCommands GetInteractiveCommands()
             => _lazyInteractiveCommands ?? throw new InvalidOperationException(EditorFeaturesResources.Engine_must_be_attached_to_an_Interactive_Window);
 
-        public void Dispose()
+        /// <summary>
+        /// Invoked on UI thread when a new language buffer is created and before it is added to the projection.
+        /// </summary>
+        private void SubmissionBufferAdded(object sender, SubmissionBufferAddedEventArgs args)
         {
-            _shutdownCancellationSource.Cancel();
-            _shutdownCancellationSource.Dispose();
-
-            _workspace.Dispose();
-            _interactiveHost.Dispose();
-
-            var interactiveWindow = _lazyInteractiveWindow;
-            if (interactiveWindow != null)
-            {
-                interactiveWindow.SubmissionBufferAdded -= SubmissionBufferAdded;
-            }
+            _threadingContext.ThrowIfNotOnUIThread();
+            _session.AddSubmissionProject(args);
         }
 
         private void CaptureClassificationSpans()
@@ -202,190 +169,6 @@ namespace Microsoft.CodeAnalysis.Editor.Interactive
             _submittedBuffers.Clear();
         }
 
-        /// <summary>
-        /// Invoked by <see cref="InteractiveHost"/> when a new process initialization completes.
-        /// </summary>
-        private void ProcessInitialized(InteractiveHostPlatformInfo platformInfo, InteractiveHostOptions options, RemoteExecutionResult result)
-        {
-            Contract.ThrowIfFalse(result.InitializationResult != null);
-
-            _ = _threadingContext.JoinableTaskFactory.RunAsync(async () =>
-            {
-                await _threadingContext.JoinableTaskFactory.SwitchToMainThreadAsync();
-                CaptureClassificationSpans();
-            });
-
-            _ = _taskQueue.ScheduleTask(nameof(ProcessInitialized), () =>
-            {
-                // clear workspace state:
-                _workspace.ClearSolution();
-                _currentSubmissionProjectId = null;
-                _lastSuccessfulSubmissionProjectId = null;
-
-                // update host state:
-                _platformInfo = platformInfo;
-                _initializationResult = result.InitializationResult;
-                UpdatePathsNoLock(result);
-
-                // Create submission projects for buffers that were added by the Interactive Window 
-                // before the process initialization completed.
-                foreach (var (buffer, languageName) in _pendingBuffers)
-                {
-                    AddSubmissionProjectNoLock(buffer, languageName);
-                }
-
-                _pendingBuffers.Clear();
-            }, _shutdownCancellationSource.Token);
-        }
-
-        private static RuntimeMetadataReferenceResolver CreateMetadataReferenceResolver(IMetadataService metadataService, InteractiveHostPlatformInfo platformInfo, ImmutableArray<string> searchPaths, string baseDirectory)
-        {
-            return new RuntimeMetadataReferenceResolver(
-                searchPaths,
-                baseDirectory,
-                gacFileResolver: platformInfo.HasGlobalAssemblyCache ? new GacFileResolver(preferredCulture: CultureInfo.CurrentCulture) : null,
-                platformAssemblyPaths: platformInfo.PlatformAssemblyPaths,
-                fileReferenceProvider: (path, properties) => metadataService.GetReference(path, properties));
-        }
-
-        private static SourceReferenceResolver CreateSourceReferenceResolver(ImmutableArray<string> searchPaths, string baseDirectory)
-            => new SourceFileResolver(searchPaths, baseDirectory);
-
-        #endregion
-
-        #region Workspace
-
-        /// <summary>
-        /// Invoked on UI thread when a new language buffer is created and before it is added to the projection.
-        /// </summary>
-        private void SubmissionBufferAdded(object sender, SubmissionBufferAddedEventArgs args)
-        {
-            _threadingContext.ThrowIfNotOnUIThread();
-
-            _taskQueue.ScheduleTask(nameof(SubmissionBufferAdded), () => AddSubmissionProjectNoLock(args.NewBuffer, _languageInfo.LanguageName), _shutdownCancellationSource.Token);
-        }
-
-        private void AddSubmissionProjectNoLock(ITextBuffer submissionBuffer, string languageName)
-        {
-            var solution = _workspace.CurrentSolution;
-            Project project;
-            var imports = ImmutableArray<string>.Empty;
-            var references = ImmutableArray<MetadataReference>.Empty;
-
-            if (_currentSubmissionProjectId == null)
-            {
-                Debug.Assert(_lastSuccessfulSubmissionProjectId == null);
-
-                // The Interactive Window may have added the first language buffer before 
-                // the host initialization has completed. Do not create a submission project 
-                // for the buffer in such case. It will be created when the initialization completes.
-                if (_initializationResult == null)
-                {
-                    _pendingBuffers.Add((submissionBuffer, languageName));
-                    return;
-                }
-
-                var initResult = _initializationResult;
-
-                imports = initResult.Imports.ToImmutableArrayOrEmpty();
-
-                var metadataService = _workspace.Services.GetRequiredService<IMetadataService>();
-                references = initResult.MetadataReferencePaths.ToImmutableArrayOrEmpty().SelectAsArray(
-                    (path, metadataService) => (MetadataReference)metadataService.GetReference(path, MetadataReferenceProperties.Assembly),
-                    metadataService);
-
-                // if a script was specified in .rps file insert a project with a document that represents it:
-                var scriptPath = initResult.ScriptPath;
-                if (scriptPath != null)
-                {
-                    project = CreateSubmissionProjectNoLock(solution, previousSubmissionProjectId: null, languageName, imports, references);
-
-                    var initDocumentId = DocumentId.CreateNewId(project.Id, debugName: scriptPath);
-                    solution = project.Solution.AddDocument(initDocumentId, Path.GetFileName(scriptPath), new FileTextLoader(scriptPath, defaultEncoding: null));
-                    _lastSuccessfulSubmissionProjectId = project.Id;
-
-                    // imports and references will be inherited:
-                    imports = ImmutableArray<string>.Empty;
-                    references = ImmutableArray<MetadataReference>.Empty;
-                }
-            }
-
-            // Project for the new submission - chain to the last submission that successfully executed.
-            project = CreateSubmissionProjectNoLock(solution, _lastSuccessfulSubmissionProjectId, languageName, imports, references);
-
-            var documentId = DocumentId.CreateNewId(project.Id, debugName: project.Name);
-            solution = project.Solution.AddDocument(documentId, project.Name, submissionBuffer.CurrentSnapshot.AsText());
-
-            _workspace.SetCurrentSolution(solution);
-
-            // opening document will start workspace listening to changes in this text container
-            _workspace.OpenDocument(documentId, submissionBuffer.AsTextContainer());
-
-            _currentSubmissionProjectId = project.Id;
-        }
-
-        private Project CreateSubmissionProjectNoLock(Solution solution, ProjectId? previousSubmissionProjectId, string languageName, ImmutableArray<string> imports, ImmutableArray<MetadataReference> references)
-        {
-            var name = "Submission#" + SubmissionCount++;
-
-            CompilationOptions compilationOptions;
-            if (previousSubmissionProjectId != null)
-            {
-                compilationOptions = solution.GetRequiredProject(previousSubmissionProjectId).CompilationOptions!;
-
-                var metadataResolver = (RuntimeMetadataReferenceResolver)compilationOptions.MetadataReferenceResolver!;
-                if (metadataResolver.PathResolver.BaseDirectory != _workingDirectory ||
-                    !metadataResolver.PathResolver.SearchPaths.SequenceEqual(_referenceSearchPaths))
-                {
-                    compilationOptions = compilationOptions.WithMetadataReferenceResolver(metadataResolver.WithRelativePathResolver(new RelativePathResolver(_referenceSearchPaths, _workingDirectory)));
-                }
-
-                var sourceResolver = (SourceFileResolver)compilationOptions.SourceReferenceResolver!;
-                if (sourceResolver.BaseDirectory != _workingDirectory ||
-                    !sourceResolver.SearchPaths.SequenceEqual(_sourceSearchPaths))
-                {
-                    compilationOptions = compilationOptions.WithSourceReferenceResolver(CreateSourceReferenceResolver(sourceResolver.SearchPaths, _workingDirectory));
-                }
-            }
-            else
-            {
-                var metadataService = _workspace.Services.GetRequiredService<IMetadataService>();
-                compilationOptions = _languageInfo.GetSubmissionCompilationOptions(
-                    name,
-                    CreateMetadataReferenceResolver(metadataService, _platformInfo, _referenceSearchPaths, _workingDirectory),
-                    CreateSourceReferenceResolver(_sourceSearchPaths, _workingDirectory),
-                    imports);
-            }
-
-            var projectId = ProjectId.CreateNewId(debugName: name);
-
-            solution = solution.AddProject(
-                ProjectInfo.Create(
-                    projectId,
-                    VersionStamp.Create(),
-                    name: name,
-                    assemblyName: name,
-                    language: languageName,
-                    compilationOptions: compilationOptions,
-                    parseOptions: _languageInfo.ParseOptions,
-                    documents: null,
-                    projectReferences: null,
-                    metadataReferences: references,
-                    hostObjectType: typeof(InteractiveScriptGlobals),
-                    isSubmission: true));
-
-            if (previousSubmissionProjectId != null)
-            {
-                solution = solution.AddProjectReference(projectId, new ProjectReference(previousSubmissionProjectId));
-            }
-
-            return solution.GetProject(projectId)!;
-        }
-
-        #endregion
-
-        #region IInteractiveEngine
-
         public bool CanExecuteCode(string text)
         {
             _threadingContext.ThrowIfNotOnUIThread();
@@ -401,11 +184,10 @@ namespace Microsoft.CodeAnalysis.Editor.Interactive
             _threadingContext.ThrowIfNotOnUIThread();
 
             var window = GetInteractiveWindow();
+
             var resetOptions = ResetOptions;
-
-            _interactiveHost.SetOutputs(window.OutputWriter, window.ErrorOutputWriter);
-
-            return ResetCoreAsync(GetHostOptions(initialize: true, resetOptions.Platform));
+            _session.Host.SetOutputs(window.OutputWriter, window.ErrorOutputWriter);
+            return _session.ResetAsync(_session.GetHostOptions(initialize: true, resetOptions.Platform));
         }
 
         /// <summary>
@@ -423,39 +205,9 @@ namespace Microsoft.CodeAnalysis.Editor.Interactive
             window.WriteLine(InteractiveEditorFeaturesResources.Resetting_execution_engine);
             window.FlushOutput();
 
-            return ResetCoreAsync(GetHostOptions(initialize, resetOptions.Platform));
-        }
-
-        public InteractiveHostOptions GetHostOptions(bool initialize, InteractiveHostPlatform? platform)
-            => InteractiveHostOptions.CreateFromDirectory(
-                _hostDirectory,
-                initialize ? _languageInfo.InteractiveResponseFileName : null,
-                CultureInfo.CurrentUICulture,
-                 platform ?? _interactiveHost.OptionsOpt?.Platform ?? InteractiveHost.DefaultPlatform);
-
-        private async Task<ExecutionResult> ResetCoreAsync(InteractiveHostOptions options)
-        {
-            try
-            {
-                _threadingContext.ThrowIfNotOnUIThread();
-
-                OnBeforeReset?.Invoke(options.Platform);
-
-                // Do not queue reset operation - invoke it directly.
-                // Code execution might be in progress when the user requests reset (via a reset button, or process terminating on its own).
-                // We need the execution to be interrupted and the process restarted, not wait for it to complete.
-
-                var result = await _interactiveHost.ResetAsync(options).ConfigureAwait(false);
-
-                // Note: Not calling UpdatePathsNoLock here. The paths will be updated by ProcessInitialized 
-                // which is executed once the new host process finishes its initialization.
-
-                return new ExecutionResult(result.Success);
-            }
-            catch (Exception e) when (FatalError.ReportAndPropagate(e))
-            {
-                throw ExceptionUtilities.Unreachable;
-            }
+            var options = _session.GetHostOptions(initialize, resetOptions.Platform);
+            OnBeforeReset?.Invoke(options.Platform);
+            return _session.ResetAsync(options);
         }
 
         /// <summary>
@@ -488,20 +240,7 @@ namespace Microsoft.CodeAnalysis.Editor.Interactive
 
                 // If process initialization is in progress we will wait with code 
                 // execution after the initialization is completed.
-
-                return await _taskQueue.ScheduleTask(nameof(ExecuteCodeAsync), async () =>
-                {
-                    var result = await _interactiveHost.ExecuteAsync(text).ConfigureAwait(false);
-                    if (result.Success)
-                    {
-                        _lastSuccessfulSubmissionProjectId = _currentSubmissionProjectId;
-
-                        // update local search paths - remote paths has already been updated
-                        UpdatePathsNoLock(result);
-                    }
-
-                    return new ExecutionResult(result.Success);
-                }, _shutdownCancellationSource.Token).ConfigureAwait(false);
+                return await _session.ExecuteCodeAsync(text).ConfigureAwait(false);
             }
             catch (Exception e) when (FatalError.ReportAndPropagate(e))
             {
@@ -520,28 +259,6 @@ namespace Microsoft.CodeAnalysis.Editor.Interactive
             return null;
         }
 
-        #endregion
-
-        private void UpdatePathsNoLock(RemoteExecutionResult result)
-        {
-            _workingDirectory = result.WorkingDirectory;
-            _referenceSearchPaths = result.ReferencePaths;
-            _sourceSearchPaths = result.SourcePaths;
-        }
-
-        public async Task SetPathsAsync(ImmutableArray<string> referenceSearchPaths, ImmutableArray<string> sourceSearchPaths, string workingDirectory)
-        {
-            try
-            {
-                var result = await _interactiveHost.SetPathsAsync(referenceSearchPaths.ToArray(), sourceSearchPaths.ToArray(), workingDirectory).ConfigureAwait(false);
-                UpdatePathsNoLock(result);
-            }
-            catch (Exception e) when (FatalError.ReportAndPropagate(e))
-            {
-                throw ExceptionUtilities.Unreachable;
-            }
-        }
-
         public string GetPrompt()
         {
             var buffer = GetInteractiveWindow().CurrentLanguageBuffer;
@@ -549,5 +266,8 @@ namespace Microsoft.CodeAnalysis.Editor.Interactive
                 ? ". "
                 : "> ";
         }
+
+        public Task SetPathsAsync(ImmutableArray<string> referenceSearchPaths, ImmutableArray<string> sourceSearchPaths, string workingDirectory)
+            => _session.SetPathsAsync(referenceSearchPaths, sourceSearchPaths, workingDirectory);
     }
 }
